fix: remove data track features
