+++ packages/language-server/src/index.ts
@@ -16,7 +16,6 @@ import { createVineDiagnosticsPlugin } from './plugins/vine-diagnostics'
 import { createDocumentHighlightForward } from './plugins/vine-document-highlight'
 import { createVineFoldingRangesPlugin } from './plugins/vine-folding-ranges'
 import { createVineTemplatePlugin } from './plugins/vine-template'
-import { track } from './track'
 
 const connection = createConnection()
 const server = createServer(connection)
@@ -89,7 +88,6 @@ connection.onInitialize(async (params) => {
           compilerOptions,
           vueCompilerOptions,
           target: 'extension',
-          track,
         },
       ),
     ]
