+++ packages/vscode-ext/package.json
@@ -97,18 +97,7 @@
     "configuration": [
       {
         "title": "Vue Vine",
-        "properties": {
-          "vue-vine.dataTrack": {
-            "type": "boolean",
-            "default": true,
-            "description": "Whether to collect event data such as extension activating, language server restarting, etc."
-          },
-          "vue-vine.hideDataTrackWarning": {
-            "type": "boolean",
-            "default": false,
-            "description": "Whether to hide the data track warning."
-          }
-        }
+        "properties": {}
       }
     ]
   },
